Разбила компоненты на подпапки в components. Поправила функцию добавления объявления в избранное.
