+++ src/components/blocks/offer/OfferGallery.tsx
