+++ src/components/blocks/offer/OfferInside.tsx
@@ -1,0 +1,20 @@
+
+export default function OfferInside() {
+  return (
+    <div className='offer__inside'>
+      <h2 className='offer__inside-title'>What&apos;s inside</h2>
+      <ul className='offer__inside-list'>
+        <li className='offer__inside-item'>Wi-Fi</li>
+        <li className='offer__inside-item'>Washing machine</li>
+        <li className='offer__inside-item'>Towels</li>
+        <li className='offer__inside-item'>Heating</li>
+        <li className='offer__inside-item'>Coffee machine</li>
+        <li className='offer__inside-item'>Baby seat</li>
+        <li className='offer__inside-item'>Kitchen</li>
+        <li className='offer__inside-item'>Dishwasher</li>
+        <li className='offer__inside-item'>Cabel TV</li>
+        <li className='offer__inside-item'>Fridge</li>
+      </ul>
+    </div>
+  );
+}
