+++ src/components/blocks/offer/OfferMarkPremium.tsx
@@ -1,0 +1,11 @@
+export default function OfferMarkPremium({ isPremium = false }: { isPremium?: boolean }) {
+  if (!isPremium) {
+    return null;
+  }
+
+  return (
+    <div className='offer__mark'>
+      <span>Premium</span>
+    </div>
+  );
+}
