+++ src/components/blocks/reviews/ReviewsForm.tsx
@@ -1,12 +1,12 @@
-import { useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
+import { useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
 
-import RatingStars from './RatingStars';
-import { useAddCommentMutation } from '../services/api';
-import Spinner from './Spinner';
+import { useAddCommentMutation } from '../../../services/api';
+import RatingStars from '../../ui/RatingStars';
+import Spinner from '../../ui/Spinner';
 
-import { ApiError } from '../types';
+import { ApiError } from '../../../types';
 
 export default function ReviewsForm() {
   const { id } = useParams();
