+++ src/components/ui/RatingStars.tsx
@@ -1,6 +1,6 @@
 
 import { Fragment } from 'react';
-import { StarRating } from '../types';
+import { StarRating } from '../../types';
 
 const ratingStars: StarRating[] = [
   { value: '5', title: 'perfect' },
