+++ src/components/ui/Spinner.tsx
@@ -1,4 +1,4 @@
-import styles from '../styles/spinner.module.css';
+import styles from '../../styles/spinner.module.css';
 
 export default function Spinner() {
   return <div className={styles.spinner}></div>;
