+++ src/layouts/LayoutMain.tsx
@@ -2,7 +2,7 @@ import { useEffect, useMemo } from 'react';
 import { shallowEqual } from 'react-redux';
 import { Outlet, useLocation } from 'react-router-dom';
 
-import Header from '../components/Header';
+import Header from '../components/layout/Header';
 import { Paths } from '../enums/paths';
 import { useTypedSelector } from '../hooks/useTypedSelector';
 import { selectActiveCity, selectAuthorizationStatus } from '../store/selectors';
